+++ status.go
@@ -1,35 +1,35 @@
 package chessongo
 
 func (g *Game) recordPosition() {
-	if g.PositionHistory == nil {
-		g.PositionHistory = map[uint64]int{}
+	if g.positionHistory == nil {
+		g.positionHistory = map[uint64]int{}
 	}
-	g.ZobristHash = g.computeZobrist()
-	g.PositionHistory[g.ZobristHash] = g.PositionHistory[g.ZobristHash] + 1
+	g.zobristHash = g.computeZobrist()
+	g.positionHistory[g.zobristHash] = g.positionHistory[g.zobristHash] + 1
 }
 
 func (g *Game) checkThreefoldRepetition() bool {
-	return g.PositionHistory != nil && g.PositionHistory[g.ZobristHash] >= 3
+	return g.positionHistory != nil && g.positionHistory[g.zobristHash] >= 3
 }
 
 func (g *Game) IsFivefoldRepetition() bool {
-	return g.PositionHistory != nil && g.PositionHistory[g.ZobristHash] >= 5
+	return g.positionHistory != nil && g.positionHistory[g.zobristHash] >= 5
 }
 
 func (g *Game) checkFiftyMoveRule() bool {
-	return g.HalfMoves >= 100
+	return g.halfMoves >= 100
 }
 
 func (g *Game) checkSeventyFiveMoveRule() bool {
-	return g.HalfMoves >= 150
+	return g.halfMoves >= 150
 }
 
 func (g *Game) refreshStatus() {
-	g.IsCheckmate = g.IsCheck && !g.hasMoves()
-	g.IsStalement = !g.IsCheckmate && !g.hasMoves()
-	g.IsMaterialDraw = g.hasInsufficientMaterial()
-	g.IsThreefoldRepetition = g.checkThreefoldRepetition()
-	g.IsFiftyMoveRule = g.checkFiftyMoveRule()
-	g.IsSeventyFiveMoveRule = g.checkSeventyFiveMoveRule()
-	g.IsFinished = g.IsCheckmate || g.IsStalement || g.IsMaterialDraw || g.IsFivefoldRepetition() || g.IsSeventyFiveMoveRule
+	g.isCheckmate = g.isCheck && !g.hasMoves()
+	g.isStalemate = !g.isCheckmate && !g.hasMoves()
+	g.isMaterialDraw = g.hasInsufficientMaterial()
+	g.isThreefoldRepetition = g.checkThreefoldRepetition()
+	g.isFiftyMoveRule = g.checkFiftyMoveRule()
+	g.isSeventyFiveMoveRule = g.checkSeventyFiveMoveRule()
+	g.isFinished = g.isCheckmate || g.isStalemate || g.isMaterialDraw || g.IsFivefoldRepetition() || g.isSeventyFiveMoveRule
 }